requirements.txt
Django
djangorestframework
psycopg2
Pillow
redis
django-redis